requirements.txt
numpy>=1.24.0
streamlit>=1.28.0
Pillow>=10.0.0
tensorflow>=2.15.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
groq>=0.4.2
python-dotenv>=1.0.0

